api/requirements.txt
Flask
Flask-Cors
opencv-python
numpy
Pillow
ultralytics